requirements.txt
aiofiles==0.8.0
binance-connector==1.15.0
fastapi==0.79.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
pydantic==1.9.1
email-validator==1.2.1
python-multipart==0.0.5
uvicorn==0.18.2
pyyaml==6.0
pydantic[email]==1.10.2
email-validator==1.0.3
websocket-client==1.4.1
